Don't remove agent when installing

+++ ext/extconf.rb
@@ -12,25 +12,23 @@ def ext_path(path)
   File.join(File.expand_path('..', __FILE__), path)
 end
 
-# Clean up possibly stale agent and lib
-FileUtils.rm_f(ext_path('appsignal-agent'))
-FileUtils.rm_f(ext_path('libappsignal.a'))
-
 arch = "#{Gem::Platform.local.cpu}-#{Gem::Platform.local.os}"
 
 if SUPPORTED_ARCHS.include?(arch)
-  archive_url = "https://#{HOST}/#{Appsignal::AGENT_VERSION}/appsignal-agent-#{arch}.tar.gz"
-  archive     = open(archive_url)
-  Gem::Package::TarReader.new(Zlib::GzipReader.open(archive)) do |tar|
-    tar.each do |entry|
-      if entry.file?
-        File.open(ext_path(entry.full_name), 'wb') do |f|
-          f.write(entry.read)
+  if !File.exists?(ext_path('appsignal-agent')) || !File.exists?(ext_path('libappsignal.a'))
+    archive_url = "https://#{HOST}/#{Appsignal::AGENT_VERSION}/appsignal-agent-#{arch}.tar.gz"
+    archive     = open(archive_url)
+    Gem::Package::TarReader.new(Zlib::GzipReader.open(archive)) do |tar|
+      tar.each do |entry|
+        if entry.file?
+          File.open(ext_path(entry.full_name), 'wb') do |f|
+            f.write(entry.read)
+          end
         end
       end
     end
+    FileUtils.chmod(0755, ext_path('appsignal-agent'))
   end
-  FileUtils.chmod(0755, ext_path('appsignal-agent'))
 
   have_library 'appsignal', 'appsignal_start'
